requirements.txt
blessings==1.6
bpython==0.16
certifi==2017.7.27.1
chardet==3.0.4
cssselect==1.0.1
curtsies==0.2.11
Django==1.11.6
django-angular==1.1.4
django-filter==1.0.4
djangorestframework==3.7.0
easy-thumbnails==2.4.2
greenlet==0.4.12
idna==2.6
lxml==4.1.0
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
Pygments==2.2.0
pyquery==1.2.17
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
wcwidth==0.1.7
